requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.8
gunicorn==20.1.0
celery==5.3.0
redis==4.6.0
django-cors-headers
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
